requirements.txt
asgiref==3.7.2
Django==4.2.2
dnspython==2.3.0
pymongo==4.4.0
python-decouple==3.8
sqlparse==0.4.4
